+++ meta_package_manager/tests/conftest.py
@@ -50,6 +50,7 @@ MANAGER_IDS = frozenset(
         "opkg",
         "pip",
         "snap",
+        "vscode",
         "yarn",
     ]
 )