+++ changelog.rst
@@ -8,6 +8,7 @@ Changelog
 .. note:: This version is not yet released and is under active development.
 
 * [mpm] Add new ``install`` command. Closes #21.
+* [vscode] Add support for Visual Studio Code plugins.
 * [mpm] Finish complete ``restore`` command implementation. Closes #38.
 * [mpm] Remove un-enforced poetry-like caret-based version specifiction from TOML
   backup files.